refactor: kafka 패키지 명 수정 및 루트 디렉토리로 이동
