+++ src/main/java/com/skyhorsemanpower/BE_AuctionPost/kafkac/KafkaConsumerConfig.java
@@ -1,4 +1,4 @@
-package com.skyhorsemanpower.BE_AuctionPost.config.kafkaconfig;
+package com.skyhorsemanpower.BE_AuctionPost.kafkac;
 
 import java.util.HashMap;
 import java.util.Map;
