+++ src/main/java/com/skyhorsemanpower/BE_AuctionPost/kafkac/KafkaProducerCluster.java
@@ -1,4 +1,4 @@
-package com.skyhorsemanpower.BE_AuctionPost.config.kafkaconfig;
+package com.skyhorsemanpower.BE_AuctionPost.kafkac;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
